#1 Fix Grid Memo
それぞれのメモをグリッド表示に変更
